requirements.txt
PyYAML
PyGithub
